Break up Export page component
